+++ core/redfishConfig.py
@@ -64,7 +64,7 @@ class RedfishConfig:
         self.dictionary['serviceversion']   = [2, '1|2         Specify the Redfish Service version. This changes command behavior based on supported schemas. Default is 2.']
         self.dictionary['showelapsed']      = [False, 'True|False  Display how long each command took. Default is False.']
         self.dictionary['trace']            = [int(TraceLevel.INFO), '4-7         Turn on additional tracing. 4=DEFAULT, 5=VERBOSE, 6=DEBUG, 7=TRACE. Default is 4.']
-        self.dictionary['urltimeout']       = [30, '<int>       How long to wait for a URL request before timing out. Default is 30.']
+        self.dictionary['urltimeout']       = [300, '<int>       How long to wait for a URL request before timing out. Default is 300.']
         self.dictionary['usefinalslash']    = [True, 'True|False  When True, all Redfish URIs will have a slash as the final character in the URL. Default is True.']
         self.dictionary['username']         = ['', '<string>    Change the username to [name] that is used to log in to the Redfish Service.']
 
@@ -145,6 +145,10 @@ class RedfishConfig:
         mcip = socket.gethostbyname(self.get_value('mcip'))
         Trace.log(TraceLevel.DEBUG, 'get_mcip() = {}'.format(mcip))
         return mcip
+
+    @classmethod
+    def get_tracelevel(self):
+        return Trace.getlevel()
 
     @classmethod
     def display(self):